fix(*): updates, cleanup and refactor after merge
