+++ frontend/src/pages/np/create/components/templates/FORRTKLReplication.tsx
@@ -2,189 +2,19 @@ import ShowOptionalWrapper from "@/components/formedible/wrappers/optional-suffi
 import ApiComboboxMultipleExpandable, {
   ApiComboboxSingle,
 } from "@/components/np/api-combobox";
-import { ResultItem, SearchEndpoint } from "@/components/np/api-endpoints";
-import { Button } from "@/components/ui/button";
 import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-} from "@/components/ui/command";
-import { Label } from "@/components/ui/label";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
+  ResultItem,
+  searchFORRTClaims,
+  WIKIDATA_ENTITY_API,
+} from "@/components/np/api-endpoints";
+import { QueryComboboxField } from "@/components/np/query-combobox";
 import { useFormedible } from "@/hooks/use-formedible";
-import ky, { KyResponse } from "ky";
-import { CheckIcon, ChevronsUpDownIcon, Loader2 } from "lucide-react";
-import { useEffect, useId, useState } from "react";
+import { useState } from "react";
 import z from "zod";
 import {
   NanopubEditorOptionFields,
   NanopubTemplateDefComponentProps,
 } from "./component-registry";
-
-// --- FORRT Claim search via SPARQL -----------------------------------------
-
-const SPARQL_ENDPOINT = "https://query.knowledgepixels.com/repo/full";
-
-async function searchFORRTClaims(term: string): Promise<ResultItem[]> {
-  if (term.length < 2) return [];
-
-  const query = `SELECT ?thing ?label WHERE {
-  graph ?g { ?thing a <https://w3id.org/sciencelive/o/terms/FORRT-Claim> }
-  OPTIONAL { graph ?g2 { ?thing <http://www.w3.org/2000/01/rdf-schema#label> ?label } }
-  FILTER(CONTAINS(LCASE(STR(?thing)), '${term.toLowerCase().replace(/'/g, "\\'")}')
-    || CONTAINS(LCASE(STR(?label)), '${term.toLowerCase().replace(/'/g, "\\'")}'))
-} LIMIT 10`;
-
-  try {
-    const res = await ky.post(SPARQL_ENDPOINT, {
-      body: new URLSearchParams({ query }),
-      headers: {
-        Accept: "application/sparql-results+xml",
-        "Content-Type": "application/x-www-form-urlencoded",
-      },
-    });
-
-    const text = await res.text();
-    const parser = new DOMParser();
-    const xmlDoc = parser.parseFromString(text, "text/xml");
-    const results = xmlDoc.getElementsByTagName("result");
-
-    return Array.from(results).map((result) => {
-      const uri =
-        result.querySelector("binding[name='thing'] uri")?.textContent || "";
-      const label =
-        result.querySelector("binding[name='label'] literal")?.textContent ||
-        uri;
-      return { uri, label };
-    });
-  } catch (e) {
-    console.error("FORRT claim search error:", e);
-    return [];
-  }
-}
-
-function FORRTClaimCombobox({
-  value,
-  onValueChange,
-}: {
-  value: ResultItem | null;
-  onValueChange: (item: ResultItem | null) => void;
-}) {
-  const id = useId();
-  const [open, setOpen] = useState(false);
-  const [inputValue, setInputValue] = useState("");
-  const [results, setResults] = useState<ResultItem[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-
-  useEffect(() => {
-    if (!inputValue || inputValue.length < 2) {
-      setResults([]);
-      return;
-    }
-    const timer = setTimeout(() => {
-      setIsLoading(true);
-      searchFORRTClaims(inputValue)
-        .then(setResults)
-        .catch(() => setResults([]))
-        .finally(() => setIsLoading(false));
-    }, 500);
-    return () => clearTimeout(timer);
-  }, [inputValue]);
-
-  return (
-    <div className="w-full space-y-2">
-      <Label htmlFor={id}>Search for a FORRT claim</Label>
-      <Popover open={open} onOpenChange={setOpen}>
-        <PopoverTrigger asChild>
-          <Button
-            id={id}
-            variant="outline"
-            role="combobox"
-            aria-expanded={open}
-            className="w-full justify-between"
-          >
-            {value ? (
-              <span className="truncate">{value.label}</span>
-            ) : (
-              <span className="text-muted-foreground/80">
-                Type to search FORRT claims...
-              </span>
-            )}
-            <ChevronsUpDownIcon className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-          </Button>
-        </PopoverTrigger>
-        <PopoverContent className="w-(--radix-popper-anchor-width) p-0">
-          <Command shouldFilter={false}>
-            <CommandInput
-              placeholder="Search FORRT claims..."
-              value={inputValue}
-              onValueChange={setInputValue}
-            />
-            <CommandList>
-              {isLoading ? (
-                <div className="flex justify-center py-6 text-center text-sm text-muted-foreground">
-                  <Loader2 className="h-4 w-4 animate-spin" />
-                </div>
-              ) : (
-                <>
-                  <CommandEmpty>
-                    {inputValue.length < 2
-                      ? "Type at least 2 characters to search..."
-                      : "No results found."}
-                  </CommandEmpty>
-                  <CommandGroup>
-                    {results.map((item) => (
-                      <CommandItem
-                        key={item.uri}
-                        value={item.uri}
-                        onSelect={() => {
-                          onValueChange(item);
-                          setOpen(false);
-                        }}
-                      >
-                        <span className="truncate">{item.label}</span>
-                        {value?.uri === item.uri && (
-                          <CheckIcon size={16} className="ml-auto" />
-                        )}
-                      </CommandItem>
-                    ))}
-                  </CommandGroup>
-                </>
-              )}
-            </CommandList>
-          </Command>
-        </PopoverContent>
-      </Popover>
-    </div>
-  );
-}
-
-// --- Wikidata search endpoint ------------------------------------------------
-
-const wikidataEndpoint: SearchEndpoint[] = [
-  {
-    name: "wikidata",
-    label: "Wikidata",
-    url: "https://www.wikidata.org/w/api.php?action=wbsearchentities&language=en&format=json&origin=*&limit=5&search=",
-    parser: async (res: KyResponse) => {
-      const json = await res.json<{
-        search: { concepturi: string; label: string; description: string }[];
-      }>();
-      return (json.search || []).map((item) => ({
-        uri: item.concepturi,
-        label: item.label,
-        description: item.description,
-      }));
-    },
-  },
-];
 
 // --- Study type options ------------------------------------------------------
 
@@ -282,12 +112,16 @@ export default function FORRTKLReplication({
         placeholder: "Search for a FORRT claim",
         required: true,
         component: ({ fieldApi }) => (
-          <FORRTClaimCombobox
+          <QueryComboboxField
             value={claimSelection}
             onValueChange={(item) => {
               setClaimSelection(item);
               fieldApi.setValue(item?.uri || "");
             }}
+            searchFunction={searchFORRTClaims}
+            labelText="Search for a FORRT claim"
+            instructionText="Search FORRT claims..."
+            placeholderText="Type to search FORRT claims..."
           />
         ),
       },
@@ -330,7 +164,7 @@ export default function FORRTKLReplication({
         },
         component: ({ fieldApi }) => (
           <ApiComboboxMultipleExpandable
-            endpoints={wikidataEndpoint}
+            endpoints={[WIKIDATA_ENTITY_API]}
             value={fieldApi.state.value || []}
             onValueChange={(items) => {
               fieldApi.setValue(items);
@@ -358,7 +192,7 @@ export default function FORRTKLReplication({
         },
         component: ({ fieldApi }) => (
           <ApiComboboxSingle
-            endpoints={wikidataEndpoint}
+            endpoints={[WIKIDATA_ENTITY_API]}
             value={fieldApi.state.value || null}
             onValueChange={(item) => {
               fieldApi.setValue(item);
